gengo: put callbacks responsible stuff in separated file
also improve logging
